requirements.txt
mlflow>=2.8.0
psutil>=5.9.0
gitpython>=3.1.40
GPUtil==1.4.0
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
numpy==2.1.2
opencv-python-headless==4.10.0.84
PyYAML==6.0.2
requests==2.32.3
tqdm==4.66.5
Pillow==10.4.0
scikit-image==0.24.0
matplotlib==3.9.2
pandas==2.2.3
